Respect our global logrotate configuration
We want to rotate all log files on a daily basis without having
individual rules per log file. Furthermore we configure every daemon which has
support for syslog to use it, instaed of logging directly to a file. This
results in a simpler logrotate configuration as everything will be
handled by the syslog daemon which requires only a SIGHUP after the log
files were moved.