Allow multiple UCARP options to be set.
To support a "preferred" UCARP master (or master switch-back) the
UCARP configuration on the preferred master requires (apart from a lower
advertisement interval, base and skew) the --preempt option to be set.
This leads to different UCARP configurations on the primary master and
secondary master nodes. Thus, we switch to a more general replacement
pattern which allows the node-configuration script to set as many
options as it needs to.